pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = ""
authors = [
    {name = "saumyapatel17",email = "saumyapatel17800@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb (>=0.6.2,<0.7.0)",
    "openai (==1.59.5)",
    "pypdf2 (==3.0.1)",
    "python-dotenv (==1.0.1)",
    "python-docx (==1.1.2)",
    "python-multipart (==0.0.20)",
    "sentence-transformers (==3.3.1)",
    "groq (==0.13.1)",
    "gradio (==5.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
